This is the commit to end all commits, hah! But seriously, this is a huge commit. I got kinda carried away and forgot a wee bit about the testing in T DD format. But I have remembered atlast and now all the tests pass. Time to create more tests.
Details for what was added is as follows:
- A user interface for interacting with the class has now been implemented. It contains a few menu options which are ~95% implemented.
-The TriangleStuffClass now contains many methods for interacting and querying it for information about its currently loaded triangle.
- NO changes to test units yet, I plan to work on that next.

Long story short there was probably about 500 lines of code difference between the last commit and this one, quite a bit was added. I realize this goes against TDD style and I will strive to fight my natural tendency to code ahead fo myself.

+++ TriangleStuffUnitTests/UnitTest1.cs
@@ -29,9 +29,9 @@ namespace TriangleStuffUnitTests
         public void Test_IsRightTriangle_ReturnsTrue()
         {
             // Arrange.
-            float AngleA = 90.0F;
-            float AngleB = 45.0F;
-            float AngleC = 45.0F;
+            double AngleA = 90.0;
+            double AngleB = 45.0;
+            double AngleC = 45.0;
             bool Result = false;
 
             // To test if the triangle is a right angled triangle it must be constructor with specific values, one of which must be 90.0.
